fix performance warnings from polars
