+++ janitor/functions/coalesce.py
@@ -5,7 +5,7 @@ from typing import Any, Optional, Union
 import pandas as pd
 import pandas_flavor as pf
 
-from janitor.functions.utils import _select_index
+from janitor.functions.select import _select_index
 from janitor.utils import check, deprecated_alias
