+++ tests/functions/test_select.py
@@ -3,7 +3,7 @@ import pandas as pd
 import pytest
 from pandas.testing import assert_frame_equal
 
-from janitor.functions.utils import DropLabel
+from janitor.functions.select import DropLabel
 
 
 @pytest.fixture
