+++ tests/functions/test_select_columns.py
@@ -8,8 +8,8 @@ import pytest
 from pandas.api.types import is_numeric_dtype
 from pandas.testing import assert_frame_equal
 
-from janitor.functions.select import get_columns
-from janitor.functions.utils import DropLabel, patterns
+from janitor.functions.select import DropLabel, get_columns
+from janitor.functions.utils import patterns
 
 
 @pytest.mark.functions
